requirements.txt
# ============================================================
# Tenantrix — Production Dependencies
# ============================================================

# Web framework
fastapi==0.135.1
uvicorn[standard]==0.41.0
python-multipart==0.0.22

# Database
SQLAlchemy==2.0.48
alembic==1.18.4
psycopg2-binary==2.9.11

# Cache & Redis
redis==5.2.0
aioredis==2.0.1

# Validation & Settings
pydantic==2.12.5
pydantic-settings==2.13.1
email-validator==2.3.0
python-dotenv==1.2.2

# Auth
PyJWT==2.11.0
passlib==1.7.4
bcrypt==3.2.2

# Rate limiting
slowapi==0.1.9
limits==5.8.0

# S3-compatible storage
boto3==1.38.44
redis==5.2.0
aioredis==2.0.1